pyproject.toml
[tool.poetry]
name = "simple-auth"
version = "0.1.0"
description = ""
authors = ["Mahdi Ghiasi <mahdiw.ghiasi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.11"
djangorestframework = "^3.13.1"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.3"
psycopg2 = "^2.9.3"
PyJWT = {extras = ["crypto"], version = "^2.3.0"}
social-auth-app-django = "^5.0.0"
djangorestframework-simplejwt = "^5.1.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
